Convert and update
Added comments and made sure everything compiled in quincy
Quincy does not work with Lambda functions and had to make a second function call for it to work right.
Also had  to make sure pow had  a (double,int) by casting and using '10.0' over '10'.
Shortened C3_P3 display to fit a Windows 7 console
Also can not initialize a vector with '={}' in quincy, so used '(7)' instead, which won't give them a default value, but isn't needed for this program.